Fix in chunk fp precision. (#1466)
* Fix in chunk fp precision.

Scaled chunks and (mantissa, exponent) -> double conversion in NewChunk treat negative exponents as multiplication by the inverse instead of division. However, division is more numerically stable in this case (i.e. a/b is more precise than a * (1/b) for some values ) leading to possible small discrepancy between h2o parse and Java standard parse of double.

For example:
System.out.println(Double.parseDouble(“-0.09375”));
System.out.println(-9375*Math.pow(10,-5));
System.out.println(-9375/Math.pow(10,5));

Produces:
-0.09375
-0.09375000000000001
-0.09375

Minor CSChunk fix. Serialization was using set8d instead of set8 for bias which is long.

Scaled chunk fix. Correctly replace NAs with imputed value in getDoubles method.

Fix in chunk compression update. Cover the case when there is only 0s and NAs (xmin is not getting updated now by 0).

Updated test, MSE improved slightly with chunk precision improvement.

Chunk compression update:
CSChunks go back to multiplication by inverse and introduced new CDecimalChunk for scale floating point values which can not be losslessly represented in any of CSchunks.

Fixed deep clone in CSChunk and CDecimalChunk

Added new CDecimal chunk to ChunkSummary

CSChunk update.

Fix in chunk fp precision.

Scaled chunks and (mantissa, exponent) -> double conversion in NewChunk treat negative exponents as multiplication by the inverse instead of division. However, division is more numerically stable in this case (i.e. a/b is more precise than a * (1/b) for some values ) leading to possible small discrepancy between h2o parse and Java standard parse of double.

For example:
System.out.println(Double.parseDouble(“-0.09375”));
System.out.println(-9375*Math.pow(10,-5));
System.out.println(-9375/Math.pow(10,5));

Produces:
-0.09375
-0.09375000000000001
-0.09375

Minor CSChunk fix. Serialization was using set8d instead of set8 for bias which is long.

Scaled chunk fix. Correctly replace NAs with imputed value in getDoubles method.

Fix in chunk compression update. Cover the case when there is only 0s and NAs (xmin is not getting updated now by 0).

Updated test, MSE improved slightly with chunk precision improvement.

Chunk compression update:
CSChunks go back to multiplication by inverse and introduced new CDecimalChunk for scale floating point values which can not be losslessly represented in any of CSchunks.

Fixed deep clone in CSChunk and CDecimalChunk

Added new CDecimal chunk to ChunkSummary

CSChunk update.

Minor CSChunk cleanup.

Removed empty ChunkPreformanceTest.

NewChunk update - fixed compression of scaled chunks.

Removed duplicate NA check in CS chunk (MichalK's review suggestion, thanks :)

Fixed set_impl(double) on scaled chunks.

minor NewChunk cleanup

NewChunk cleanup - removed duplicate code.

Updated GBM tests to pass with improved chunk precision - couple of test had comparisons against hard coded values which failed with new  ( more accurate) chunk values.
For lack of better options, I just added tolerance to the comparison so that the test pass.
You can verify the change in gbm results is solely due to chunk change by turning setDecimal in CSChunk to no-op.

Fixed new chunk test to use more precise division instead of multiplication by inverse.

CS chunk fix - hasFloat is true if isDecimal.

Minor cleanup in AstMerge - extracting values from Chunk to NewChunk should be done via appropriate method instead of atd/atUUID/...

* Relaxed comparison tolerance in test - was not passing with 1e-12 anymore due to a chunk precision update.

Updated python test to pass with updated CSChunks - the difference in gbm model/lift gain table de to improved precision of CSChunks - can be verified by turning CSChunk::setDecimal to a no-op. Updated numbers so that the test passes.

+++ h2o-core/src/main/java/water/fvec/C4Chunk.java
@@ -6,7 +6,7 @@ import water.util.UnsafeUtils;
  * The empty-compression function, where data is in 'int's.
  */
 public class C4Chunk extends Chunk {
-  static protected final long _NA = Integer.MIN_VALUE;
+  static protected final int _NA = Integer.MIN_VALUE;
   C4Chunk( byte[] bs ) { _mem=bs; _start = -1; set_len(_mem.length>>2); }
   @Override protected final long at8_impl( int i ) {
     long res = UnsafeUtils.get4(_mem,i<<2);
